requirements.txt
#Python 3.13.2

#Requirements are poorly set out currently (sorry)

# Core dependencies
numpy>=1.24.0 #possibly not needed
Pillow>=11.1.0
PyQt6>=6.8.1
dependency-injector>=4.41.0

# Image processing and machine learning
scipy>=1.15.2  # For FFT and signal processing
scikit-learn>=1.6.1  # For clustering algorithms

# Development tools // not yet checked
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0


#Wordcloud